Remove MPMC individual benchmark
The threads spawn when launching the benchmark are not scheduled
during the whole benchmark period. Thus, no item is pop and we have
really low empty check time, which is not accurate for comparison.

Even though it works, the rate of the pop depends on how fast we can
provide new data, which is limited by the push (producer) rate.
Therefore, we still need to adjust the ratio between the producer and
consumers to accurately measure the pop rate.

Maybe we can run a group of experiments that adjust producer and
consumer roles based on the index. For example, pivot = 1..16,
and index < pivot are producers and index >= pivot are consumers.
That is to say, we add a new dimension in our final experiments
for comparison.